+++ lib/slack/real_time/stores/base.rb
@@ -1,11 +1,23 @@
 # frozen_string_literal: true
+
 module Slack
   module RealTime
     module Stores
       # Doesn't store anything.
       class Base
+        @events = Hash.new { |h, k| h[k] = [] }
+
         class << self
-          attr_accessor :events
+          attr_reader :events
+
+          def inherited(subclass)
+            super
+            subclass.instance_variable_set :@events, events.dup
+          end
+
+          def on(event, &handler)
+            events[event.to_s] << handler
+          end
         end
 
         attr_accessor :users, :bots, :channels, :groups, :teams, :ims
@@ -19,12 +31,6 @@ module Slack
         end
 
         def initialize(_attrs); end
-
-        def self.on(event, &block)
-          self.events ||= {}
-          self.events[event.to_s] ||= []
-          self.events[event.to_s] << block
-        end
       end
     end
   end